Register for intent ACTION_BOOT_COMPLETED
We are currently listening for LOCKED_BOOT_COMPLETED, but we
should be listening for ACTION_BOOT_COMPLETED:

Upon receipt of this broadcast, the user is unlocked and both
device-protected and credential-protected storage can accessed safely.

This will let us clean up the SharedPreferences when Developer Options
are disabled (or never have been enabled) without hitting this exception.

Bug: 129445124
Test: atest CtsAngleIntegrationHostTestCases
Change-Id: Icbef57bc73cc0930a14ac2bd9afaf43d42bc6949