Fix com.android.settings.summary to use android:resource
From the Android documentation:

android:resource
A reference to a resource.
android:value
The value assigned to the item.

It seems that android:resource can be looked up more correctly so the
correct string can be used based on what the current locale is, while
android:value appears to be a bit more stuck.

We want android:resource.

Bug: 132049074
Test: Set locale to es-US and verify the text is correct.
Change-Id: If2d77c99412d0a5261d70b979a99bc9c74f184b9