Remove hidden API usage in ANGLE
ANGLE is a Mainline Module, and to achieve SSI(single system image),
apps on /product have to not use hidden API, and it will be enforced.
In this case, most of the violations are from SettingsLib, not app
itself, but it looks it doesn't need SettingsLib.

Bug: 132749970
Test: atest CtsAngleIntegrationHostTestCases
Test: art/tools/veridex/appcompat.sh --dex-file=out/target/product/blueline/product/priv-app/GoogleANGLE/GoogleANGLE.apk
Test: Manually verify ANGLE can be searched for in Settings
Change-Id: I42905e87233ea90e0f63763e71491ccdf0b33591
Merged-In: I42905e87233ea90e0f63763e71491ccdf0b33591